add routing for requests to /api/reviews
