+++ server/server-config.js
@@ -1,0 +1,6 @@
+const { TOKEN } = require('./config');
+
+module.exports = {
+  baseUrl: 'https://app-hrsei-api.herokuapp.com/api/fec2/hr-sfo',
+  authorization: TOKEN,
+};
